Add apex pubkey injection routine for repacking

+++ aosp_apex_injector.py
@@ -79,7 +79,12 @@ def repackage_apex_file(aosp_path, apex_file_path, output_file_path, lunch_targe
                                                                                   output_file_path,
                                                                                   lunch_target,
                                                                                   canned_fs_config)
-                inject_apex_avb_public_key_module(apex_file_path, avb_pub_key_path)
+                is_success, log_message = inject_apex_avb_public_key_module(apex_file_path, avb_pub_key_path)
+                if is_success:
+                    log_message = f"APEX extraction success: {output_file_path}"
+                    logging.info(log_message)
+                else:
+                    log_message = f"APEX extraction failed. {output_file_path} | {log_message}"
         else:
             log_message = f"APEX extraction failed. {apex_file_path} | {log_message}"
     except Exception as e:
@@ -87,44 +92,51 @@ def repackage_apex_file(aosp_path, apex_file_path, output_file_path, lunch_targe
     return is_success, log_message
 
 def inject_apex_avb_public_key_module(input_apex, avb_pub_key_path):
-    is_success = False
-    log_message = ""
+    logging.info(f"Injecting AVB public key in APEX module for repacker: {input_apex}")
     apex_main_folder = os.path.dirname(input_apex)
     key_name = os.path.basename(avb_pub_key_path)
-    copy_avb_public_key_to_apex_module(input_apex, apex_main_folder, avb_pub_key_path)
-    android_bp_file = os.path.join(apex_main_folder, "Android.bp")
-    if os.path.exists(android_bp_file):
-        with open(android_bp_file, 'r+') as android_bp:
-            content = android_bp.read()
-            if "apex_key" not in content:
-                insert_position = content.find('name:')
-                if insert_position != -1:
-                    content = content[:insert_position] + f"apex_key: {{public_key: \"{key_name}\",}},\n" + content[insert_position:]
-                    android_bp.seek(0)
-                    android_bp.write(content)
-                    android_bp.truncate()
-                    is_success = True
-                    logging.info(f"AVB public key injected in APEX module Android.bp file: {android_bp_file}")
+    is_success, log_message = copy_avb_public_key_to_apex_module(input_apex, apex_main_folder, avb_pub_key_path)
+    if is_success:
+        android_bp_file = os.path.join(apex_main_folder, "Android.bp")
+        if os.path.exists(android_bp_file):
+            with open(android_bp_file, 'r+') as android_bp:
+                content = android_bp.read()
+                if "apex_key" not in content:
+                    insert_position = content.find('name:')
+                    if insert_position != -1:
+                        content = content[:insert_position] + f"apex_key: {{public_key: \"{key_name}\",}},\n" + content[insert_position:]
+                        android_bp.seek(0)
+                        android_bp.write(content)
+                        android_bp.truncate()
+                        is_success = True
+                        logging.info(f"AVB public key injected in APEX module Android.bp file: {android_bp_file}")
+                    else:
+                        logging.error(f"Error injecting AVB public key in APEX module: {android_bp_file}")
+                        is_success = False
                 else:
-                    logging.error(f"Error injecting AVB public key in APEX module: {android_bp_file}")
-                    is_success = False
-            else:
-                logging.info(f"AVB public key already injected in APEX module: {android_bp_file}")
-                is_success = True
+                    logging.info(f"AVB public key already injected in APEX module: {android_bp_file}")
+                    is_success = True
+    else:
+        logging.error(f"Error copying AVB public key to APEX module: {log_message}")
     return is_success, log_message
 
 def copy_avb_public_key_to_apex_module(input_apex, apex_main_folder, avb_pub_key_path):
+    is_success = False
+    log_message = ""
     apex_name = os.path.basename(input_apex).replace(".apex", "")
     public_key_name = f"{apex_name}_pubkey.pem"
     public_key_out = os.path.join(apex_main_folder, public_key_name)
     if not os.path.exists(public_key_out):
-        logging.info(f"Copying AVB public key to APEX module: {public_key_out}")
+        log_message = f"Copying AVB public key to APEX module: {public_key_out}"
+        logging.error(log_message)
     elif not os.path.exists(avb_pub_key_path):
-        logging.info(f"AVB public key already exists in APEX module: {public_key_out}")
+        log_message = f"AVB public key already exists in APEX module: {public_key_out}"
+        logging.error(log_message)
     else:
         shutil.copy(avb_pub_key_path, public_key_out)
         logging.info(f"AVB public key copied to APEX build module: {public_key_out}")
-
+        is_success = True
+    return is_success, log_message
 
 
 def get_apex_build_intermediate_folder(target_out_path):
